# docs/controlled/API-0001-openapi.yaml
openapi: 3.0.3
info:
  title: FLOTA API
  version: "1.0.0"
  description: |
    API de Control Flota (FLOTA). Documento controlado API-0001.
    Autenticación por JWT Bearer; tiempo real por Socket.IO (ver API-0001-README.md).
    Respuestas con forma { ok: true, ... } o { ok: false, error }.
  contact:
    name: SIMARP
servers:
  - url: https://flota.simarp.net
    description: Producción
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/admin/trips:
    get:
      tags: [admin]
      summary: Reporte de viajes con filtros
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { type: string } }
        - { name: driver_id, in: query, schema: { type: integer } }
        - { name: company_id, in: query, schema: { type: integer } }
      responses: { '200': { description: OK } }
...
  /api/admin/billing:
    get:
      tags: [admin]
      summary: Resumen de facturación por empresa (servicios completados)
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  rows:
                    type: array
                    items:
                      type: object
                      properties:
                        company_id: { type: integer, nullable: true }
                        company_name: { type: string }
                        services: { type: integer }
                        total: { type: integer }
                        paid: { type: integer }
                        pending: { type: integer }
                        pending_count: { type: integer }
                  totals: { type: object }
  /api/admin/trip_billing:
    post:
      tags: [admin]
      summary: Marcar estado de facturación de un servicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trip_id, billing_status]
              properties:
                trip_id: { type: integer }
                billing_status: { type: string, enum: [pending, paid, void] }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Ok:
      type: object
      properties: { ok: { type: boolean, example: true } }
...